// fichain-web/src/proto/call_data_pb.ts
// @generated by protoc-gen-es v2.5.1 with parameter "target=ts"
// @generated from file call_data.proto (package call_data, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file call_data.proto.
 */
export const file_call_data: GenFile = /*@__PURE__*/
  fileDesc("Cg9jYWxsX2RhdGEucHJvdG8SCWNhbGxfZGF0YSI8ChlDYWxsU21hcnRDb250cmFjdEhhc2hEYXRhEhEKCVRvQWRkcmVzcxgBIAEoDBIMCgREYXRhGAIgASgMIkYKFUNhbGxTbWFydENvbnRyYWN0RGF0YRIRCglUb0FkZHJlc3MYASABKAwSDAoERGF0YRgCIAEoDBIMCgRTaWduGAMgASgMIjcKGUNhbGxTbWFydENvbnRyYWN0UmVzcG9uc2USDAoESGFzaBgBIAEoDBIMCgREYXRhGAIgASgMQghaBi9wcm90b2IGcHJvdG8z");

/**
 * @generated from message call_data.CallSmartContractHashData
 */
export type CallSmartContractHashData = Message<"call_data.CallSmartContractHashData"> & {
  /**
   * @generated from field: bytes ToAddress = 1;
   */
  ToAddress: Uint8Array;

  /**
   * @generated from field: bytes Data = 2;
   */
  Data: Uint8Array;
};

/**
 * Describes the message call_data.CallSmartContractHashData.
 * Use `create(CallSmartContractHashDataSchema)` to create a new message.
 */
export const CallSmartContractHashDataSchema: GenMessage<CallSmartContractHashData> = /*@__PURE__*/
  messageDesc(file_call_data, 0);

/**
 * @generated from message call_data.CallSmartContractData
 */
export type CallSmartContractData = Message<"call_data.CallSmartContractData"> & {
  /**
   * @generated from field: bytes ToAddress = 1;
   */
  ToAddress: Uint8Array;

  /**
   * @generated from field: bytes Data = 2;
   */
  Data: Uint8Array;

  /**
   * @generated from field: bytes Sign = 3;
   */
  Sign: Uint8Array;
};

/**
 * Describes the message call_data.CallSmartContractData.
 * Use `create(CallSmartContractDataSchema)` to create a new message.
 */
export const CallSmartContractDataSchema: GenMessage<CallSmartContractData> = /*@__PURE__*/
  messageDesc(file_call_data, 1);

/**
 * @generated from message call_data.CallSmartContractResponse
 */
export type CallSmartContractResponse = Message<"call_data.CallSmartContractResponse"> & {
  /**
   * @generated from field: bytes Hash = 1;
   */
  Hash: Uint8Array;

  /**
   * @generated from field: bytes Data = 2;
   */
  Data: Uint8Array;
};

/**
 * Describes the message call_data.CallSmartContractResponse.
 * Use `create(CallSmartContractResponseSchema)` to create a new message.
 */
export const CallSmartContractResponseSchema: GenMessage<CallSmartContractResponse> = /*@__PURE__*/
  messageDesc(file_call_data, 2);